@@ -1,74 +1,64 @@
 {
   "name": "tabliss",
-  "version": "1.18.2",
+  "version": "2.0.0-beta.1",
   "license": "GPL-3.0",
   "private": true,
   "scripts": {
     "build": "npm run-script build:web",
-    "build:chrome": "NODE_ENV=production BUILD_TARGET=chrome webpack",
+    "build:chromium": "NODE_ENV=production BUILD_TARGET=chromium webpack",
     "build:firefox": "NODE_ENV=production BUILD_TARGET=firefox webpack",
     "build:web": "NODE_ENV=production BUILD_TARGET=web webpack",
-    "lint": "tslint --project tsconfig.json",
     "start": "webpack-dev-server",
     "test": "jest --config jest.json",
     "translations": "node ./scripts/translations.js"
   },
   "dependencies": {
-    "feather-icons": "^4.7.3",
-    "in-browser-language": "^1.0.2",
+    "@sentry/browser": "^5.5.0",
+    "date-fns": "^2.0.0-beta.2",
+    "date-fns-tz": "^1.0.7",
+    "feather-icons": "^4.22.1",
+    "in-browser-language": "^1.0.3",
     "localforage": "^1.7.1",
-    "lodash-es": "^4.17.10",
-    "luxon": "^1.0.0",
     "normalize.css": "^8.0.0",
     "parse-link-header": "^1.0.1",
-    "raven-js": "^3.25.2",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
-    "react-error-boundary": "^1.2.3",
-    "react-intl": "^2.4.0",
-    "react-redux": "^5.0.7",
-    "react-transition-group": "^2.3.1",
-    "redux": "^3.7.2",
-    "redux-persist": "^4.10.2",
-    "redux-persist-transform-filter": "0.0.16",
-    "screenfull": "^3.3.3",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-error-boundary": "^1.2.5",
+    "react-intl": "^3.0.0-beta.18",
+    "react-redux": "^7.1.0",
+    "redux": "^4.0.1",
+    "redux-persist": "^6.0.0-pre2.1",
+    "redux-persist-webextension-storage": "^1.0.2",
     "tlds": "^1.203.1",
     "uuid": "^3.2.1"
   },
   "devDependencies": {
-    "@types/chrome": "^0.0.66",
-    "@types/jest": "^22.2.3",
-    "@types/lodash-es": "^4.17.0",
-    "@types/luxon": "^0.5.2",
+    "@types/chrome": "0.0.86",
+    "@types/jest": "^24.0.15",
     "@types/node": "^10.1.4",
     "@types/parse-link-header": "^1.0.0",
-    "@types/react": "^16.3.14",
-    "@types/react-dom": "^16.0.5",
-    "@types/react-intl": "^2.3.8",
-    "@types/react-redux": "^5.0.20",
-    "@types/react-transition-group": "^2.0.9",
-    "@types/screenfull": "^3.3.0",
+    "@types/react": "^16.8.19",
+    "@types/react-dom": "^16.8.4",
+    "@types/react-redux": "^7.1.0",
+    "@types/tlds": "^1.199.0",
     "@types/uuid": "^3.4.3",
     "clean-webpack-plugin": "^0.1.19",
     "copy-webpack-plugin": "^4.5.1",
-    "css-loader": "^0.28.11",
+    "css-loader": "^3.0.0",
     "dotenv": "^5.0.0",
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.0.0",
-    "jest": "^23.0.0",
-    "mini-css-extract-plugin": "^0.4.0",
+    "jest": "^24.8.0",
+    "mini-css-extract-plugin": "^0.8.0",
     "node-sass": "^4.12.0",
+    "prettier": "1.17.1",
     "raw-loader": "^0.5.1",
     "react-intl-translations-manager": "^5.0.3",
     "sass-loader": "^7.0.0",
-    "source-map-loader": "^0.2.3",
     "sw-precache-webpack-plugin": "^0.11.5",
-    "ts-jest": "^22.4.6",
+    "ts-jest": "^24.0.2",
     "ts-loader": "^4.0.0",
-    "tslint": "^5.10.0",
-    "tslint-loader": "^3.6.0",
-    "tslint-react": "^3.6.0",
-    "typescript": "^2.8.4",
+    "typescript": "^3.5.2",
     "typescript-react-intl": "^0.2.5",
     "url-loader": "^1.0.0",
     "webpack": "^4.14.0",
