@@ -1,58 +1,50 @@
-import * as React from 'react';
-const categories: Category[] = require('./categories.json');
+import React, { FC } from 'react';
 
-interface Category {
-  key: string;
-  name: string;
-}
+import categories from './categories';
+import { Props, defaultData } from './types';
 
-interface Props {
-  category?: string;
-  onChange: (settings: { [key: string]: string|undefined }) => void;
-}
+const QuoteSettings: FC<Props> = ({ data = defaultData, setData }) => (
+  <div className="QuoteSettings">
+    <label>
+      <input
+        type="radio"
+        checked={data.category === undefined}
+        onChange={() => setData({ category: undefined })}
+      />{' '}
+      All Categories
+    </label>
 
-class QuoteSettings extends React.PureComponent<Props> {
-  render() {
-    return (
-      <div className="QuoteSettings">
-        <label>
-          <input
-            type="radio"
-            checked={this.props.category === undefined}
-            onChange={() => this.selectCategory(undefined)}
-          />
-          {' '}
-          All Categories
-        </label>
+    {categories.map(category => (
+      <label key={category.key}>
+        <input
+          type="radio"
+          checked={data.category === category.key}
+          onChange={() => setData({ category: category.key })}
+        />{' '}
+        {category.name}
+      </label>
+    ))}
 
-        {categories.map(category =>
-          <label key={category.key}>
-            <input
-              type="radio"
-              checked={this.props.category === category.key}
-              onChange={() => this.selectCategory(category.key)}
-            />
-            {' '}
-            {category.name}
-          </label>
-        )}
-
-        <p>
-          Powered by
-          {' '}
-          <a href="https://theysaidso.com" target="_blank" rel="noopener noreferrer">They Said So</a>
-          {' '}
-          and
-          {' '}
-          <a href="http://www.developerexcuses.com" target="_blank" rel="noopener noreferrer">Developer Excuses</a>.
-        </p>
-      </div>
-    );
-  }
-
-  private selectCategory(category?: string) {
-    this.props.onChange({ category });
-  }
-}
+    <p>
+      Powered by{' '}
+      <a
+        href="https://theysaidso.com/"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        They Said So
+      </a>
+      {' and '}
+      <a
+        href="http://www.developerexcuses.com/"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        Developer Excuses
+      </a>
+      .
+    </p>
+  </div>
+);
 
 export default QuoteSettings;
