@@ -1,9 +1,7 @@
-import * as React from 'react';
+import React from 'react';
 import { render } from 'react-dom';
-import { Provider as StoreProvider } from 'react-redux';
-import { App } from './app';
-import { store } from './data';
-import { IntlProvider } from './locales';
+
+import Root from './Root';
 import { register as registerErrorHandler } from './errorHandler';
 import { register as registerServiceWorker } from './serviceWorker';
 
@@ -13,16 +11,12 @@ if (process.env.NODE_ENV === 'production') {
 }
 
 // Render app into root element
-render(
-  <StoreProvider store={store}>
-    <IntlProvider>
-      <App />
-    </IntlProvider>
-  </StoreProvider>,
-  document.getElementById('root'),
-);
+render(<Root />, document.getElementById('root'));
 
 // Register service worker on web
-if (process.env.NODE_ENV === 'production' && process.env.BUILD_TARGET === 'web') {
+if (
+  process.env.NODE_ENV === 'production' &&
+  process.env.BUILD_TARGET === 'web'
+) {
   registerServiceWorker();
 }
