@@ -1,54 +1,36 @@
-import * as React from 'react';
-import { Settings } from '../../interfaces';
+import React, { FC } from 'react';
 
-interface Props {
-  expand?: boolean;
-  nsfw?: boolean;
-  tag?: string;
-  onChange: (settings: Settings) => void;
-}
+import { Props, defaultData } from './types';
 
-class GiphySettings extends React.PureComponent<Props> {
-  static defaultProps = {
-    expand: false,
-    tag: 'pattern',
-    nsfw: false,
-  };
+const GiphySettings: FC<Props> = ({ data = defaultData, setData }) => (
+  <div className="GiphySettings">
+    <label>
+      Tag
+      <input
+        type="text"
+        value={data.tag}
+        onChange={event => setData({ ...data, tag: event.target.value })}
+      />
+    </label>
 
-  render() {
-    return (
-      <div className="GiphySettings">
-        <label>
-          Tag
-          <input
-            type="text"
-            value={this.props.tag}
-            onChange={event => this.props.onChange({ tag: event.target.value })}
-          />
-        </label>
+    <label>
+      <input
+        type="checkbox"
+        checked={data.nsfw}
+        onChange={event => setData({ ...data, nsfw: !data.nsfw })}
+      />{' '}
+      Allow NSFW
+    </label>
 
-        <label>
-          <input
-            type="checkbox"
-            checked={this.props.nsfw}
-            onChange={event => this.props.onChange({ nsfw: ! this.props.nsfw })}
-          />
-          {' '}
-          Allow NSFW
-        </label>
-
-        <label>
-          <input
-            type="checkbox"
-            checked={this.props.expand}
-            onChange={event => this.props.onChange({ expand: ! this.props.expand })}
-          />
-          {' '}
-          Stretch to fill screen
-        </label>
-      </div>
-    );
-  }
-}
+    <label>
+      <input
+        type="checkbox"
+        checked={data.expand}
+        onChange={event => setData({ ...data, expand: !data.expand })}
+      />{' '}
+      Stretch to fill screen
+    </label>
+  </div>
+);
 
 export default GiphySettings;
