@@ -1,25 +1,16 @@
-import * as React from 'react';
+import React, { FC } from 'react';
 
-interface Props {
-  angle?: number;
-  from?: string;
-  to?: string;
-  type?: string;
-}
+import { Props, defaultData } from './types';
 
-class Gradient extends React.PureComponent<Props> {
-  static defaultProps: Props = {
-    angle: 0,
-    from : '#3498db',
-    to: '#9b59b6',
-    type: 'linear-gradient',
-  };
-
-  render() {
-    const backgroundImage = `${this.props.type}(${this.props.angle}deg, ${this.props.from}, ${this.props.to})`;
-
-    return <div className="Gradient fullscreen" style={{ backgroundImage }} />;
-  }
-}
+const Gradient: FC<Props> = ({ data = defaultData }) => (
+  <div
+    className="Gradient fullscreen"
+    style={{
+      backgroundImage: `${data.type}(${data.angle}deg, ${data.from}, ${
+        data.to
+      })`,
+    }}
+  />
+);
 
 export default Gradient;
