@@ -1,2 +1,13 @@
-export { default as Gradient } from './Gradient';
-export { default as GradientSettings } from './GradientSettings';
+import { Config } from '../../types';
+import Gradient from './Gradient';
+import GradientSettings from './GradientSettings';
+
+const config: Config = {
+  key: 'background/gradient',
+  name: 'Colour Gradient',
+  description: 'Add more splashes of colour.',
+  dashboardComponent: Gradient,
+  settingsComponent: GradientSettings,
+};
+
+export default config;
