@@ -1,6 +1,2 @@
-export * from './interfaces';
-export * from './registry';
-
-// Register plugins
-import './backgrounds/register';
-import './widgets/register';
+export * from './plugins';
+export * from './types';
