@@ -1,30 +1,24 @@
-import * as React from 'react';
-import { Settings } from '../../interfaces';
+import React, { FC } from 'react';
 
-interface Props {
-  input?: string;
-  onChange: (settings: Settings) => void;
-}
+import { Props, defaultData } from './types';
 
-const CssSettings: React.StatelessComponent<Props> = ({ input = '', onChange }) => {
-  return (
-    <div className="CssSettings">
-      <label>
-        CSS Snippet
-        <textarea
-          rows={3}
-          style={{ fontFamily: 'monospace' }}
-          value={input}
-          onChange={event => onChange({ input: event.target.value })}
-        />
-      </label>
+const CssSettings: FC<Props> = ({ data = defaultData, setData }) => (
+  <div className="CssSettings">
+    <label>
+      CSS Snippet
+      <textarea
+        rows={3}
+        style={{ fontFamily: 'monospace' }}
+        value={data.input}
+        onChange={event => setData({ input: event.target.value })}
+      />
+    </label>
 
-      <p className="info">
-        Warning: this functionality is intended for advanced users.
-        Custom styles may break at any time.
-      </p>
-    </div>
-  );
-};
+    <p className="info">
+      Warning: this functionality is intended for advanced users. Custom styles
+      may break at any time.
+    </p>
+  </div>
+);
 
 export default CssSettings;
