@@ -1,2 +1,13 @@
-export { default as Css } from './Css';
-export { default as CssSettings } from './CssSettings';
+import { Config } from '../../types';
+import Css from './Css';
+import CssSettings from './CssSettings';
+
+const config: Config = {
+  key: 'widget/css',
+  name: 'Custom CSS',
+  description: 'Make your new tab more style-ish (advanced users).',
+  dashboardComponent: Css,
+  settingsComponent: CssSettings,
+};
+
+export default config;
