@@ -1,30 +1,18 @@
-import * as React from 'react';
-import { Settings } from '../../interfaces';
+import React, { FC } from 'react';
 
-interface Props {
-  name?: string;
-  onChange: (settings: Settings) => void;
-}
+import { Props, defaultData } from './types';
 
-class GreetingSettings extends React.PureComponent<Props> {
-  static defaultProps = {
-    name: '',
-  };
-
-  render() {
-    return (
-      <div className="GreetingSettings">
-        <label>
-          Name
-          <input
-            type="text"
-            value={this.props.name}
-            onChange={event => this.props.onChange({ name: event.target.value })}
-          />
-        </label>
-      </div>
-    );
-  }
-}
+const GreetingSettings: FC<Props> = ({ data = defaultData, setData }) => (
+  <div className="GreetingSettings">
+    <label>
+      Name
+      <input
+        type="text"
+        value={data.name}
+        onChange={event => setData({ name: event.target.value })}
+      />
+    </label>
+  </div>
+);
 
 export default GreetingSettings;
