@@ -1,47 +1,23 @@
-import * as React from 'react';
-import debounce from 'lodash-es/debounce';
+import { FC, useEffect } from 'react';
 
-interface Props {
-  input?: string;
-}
+import { Props, defaultData } from './types';
 
-class Js extends React.PureComponent<Props> {
-  private debouncedAttach = debounce(this.attach, 500);
-
-  componentDidMount() {
-    this.attach();
-  }
-
-  componentDidUpdate() {
-    this.detach();
-    this.debouncedAttach();
-  }
-
-  componentWillUnmount() {
-    this.detach();
-  }
-
-  render() {
-    return null;
-  }
-
-  private detach() {
-    const script = document.getElementById('CustomJs');
-
-    if (script) {
-      document.head.removeChild(script);
-    }
-  }
-
-  private attach() {
+const Js: FC<Props> = ({ data = defaultData }) => {
+  useEffect(() => {
     const script = document.createElement('script');
 
     script.id = 'CustomJs';
     script.type = 'text/javascript';
-    script.appendChild(document.createTextNode(this.props.input || ''));
+    script.appendChild(document.createTextNode(data.input));
 
     document.head.appendChild(script);
-  }
-}
+
+    return () => {
+      document.head.removeChild(script);
+    };
+  }, [data.input]);
+
+  return null;
+};
 
 export default Js;
