@@ -1,2 +1,13 @@
-export { default as Js } from './Js';
-export { default as JsSettings } from './JsSettings';
+import { Config } from '../../types';
+import Js from './Js';
+import JsSettings from './JsSettings';
+
+const config: Config = {
+  key: 'widget/js',
+  name: 'Custom JS',
+  description: 'Program in your program.',
+  dashboardComponent: Js,
+  settingsComponent: JsSettings,
+};
+
+export default config;
