@@ -1,2 +1,13 @@
-export { default as Links } from './Links';
-export { default as LinksSettings} from './LinksSettings';
+import { Config } from '../../types';
+import Links from './Links';
+import LinksSettings from './LinksSettings';
+
+const config: Config = {
+  key: 'widget/links',
+  name: 'Quick Links',
+  description: 'I heard you like bookmarks.',
+  dashboardComponent: Links,
+  settingsComponent: LinksSettings,
+};
+
+export default config;
