@@ -1,11 +1,11 @@
-import * as React from 'react';
+import React, { FC } from 'react';
 
-const Message = ({ message = 'Add something witty' }) => {
-  return (
-    <div className="Message">
-      <h3 style={{ whiteSpace: 'pre' }}>{message}</h3>
-    </div>
-  );
-};
+import { Props, defaultData } from './types';
+
+const Message: FC<Props> = ({ data = defaultData }) => (
+  <div className="Message">
+    <h3 style={{ whiteSpace: 'pre' }}>{data.messages[0]}</h3>
+  </div>
+);
 
 export default Message;
