@@ -1,96 +1,63 @@
-import * as React from 'react';
-import { Settings } from '../../interfaces';
+import React, { FC } from 'react';
 
-interface Props {
-  mode: string;
-  hour12: boolean;
-  showDate: boolean;
-  showMinutes: boolean;
-  showSeconds: boolean;
-  timezone?: string;
-  onChange: (settings: Settings) => void;
-}
+import { Props, defaultData } from './types';
 
-class TimeSettings extends React.PureComponent<Props> {
-  static defaultProps: Partial<Props> = {
-    mode: 'digital',
-    hour12: false,
-    showDate: false,
-    showMinutes: true,
-    showSeconds: false,
-  };
+const TimeSettings: FC<Props> = ({ data = defaultData, setData }) => (
+  <div className="TimeSettings">
+    <label>
+      <input
+        type="radio"
+        checked={data.mode === 'analogue'}
+        onChange={() => setData({ ...data, mode: 'analogue' })}
+      />{' '}
+      Analogue
+    </label>
 
-  render() {
-    return (
-      <div className="TimeSettings">
-        <label>
-          <input
-            type="radio"
-            checked={this.props.mode === 'analogue'}
-            onChange={() => this.props.onChange({ mode: 'analogue' })}
-          />
-          {' '}
-          Analogue
-        </label>
+    <label>
+      <input
+        type="radio"
+        checked={data.mode === 'digital' && data.hour12}
+        onChange={() => setData({ ...data, mode: 'digital', hour12: true })}
+      />{' '}
+      12-hour digital
+    </label>
 
-        <label>
-          <input
-            type="radio"
-            checked={this.props.mode === 'digital' && this.props.hour12}
-            onChange={() => this.props.onChange({
-              mode: 'digital',
-              hour12: true,
-            })}
-          />
-          {' '}
-          12-hour digital
-        </label>
+    <label>
+      <input
+        type="radio"
+        checked={data.mode === 'digital' && !data.hour12}
+        onChange={() => setData({ ...data, mode: 'digital', hour12: false })}
+      />{' '}
+      24-hour digital
+    </label>
 
-        <label>
-          <input
-            type="radio"
-            checked={this.props.mode === 'digital' && ! this.props.hour12}
-            onChange={() => this.props.onChange({
-              mode: 'digital',
-              hour12: false,
-            })}
-          />
-          {' '}
-          24-hour digital
-        </label>
+    <label>
+      <input
+        type="checkbox"
+        checked={data.showSeconds}
+        onChange={() => setData({ ...data, showSeconds: !data.showSeconds })}
+      />{' '}
+      Display seconds
+    </label>
 
-        <label>
-          <input
-            type="checkbox"
-            checked={this.props.showSeconds}
-            onChange={() => this.props.onChange({ showSeconds: ! this.props.showSeconds })}
-          />
-          {' '}
-          Display seconds
-        </label>
+    <label>
+      <input
+        type="checkbox"
+        checked={data.showMinutes}
+        onChange={() => setData({ ...data, showMinutes: !data.showMinutes })}
+      />{' '}
+      Display minutes
+    </label>
 
-        <label>
-          <input
-            type="checkbox"
-            checked={this.props.showMinutes}
-            onChange={() => this.props.onChange({ showMinutes: ! this.props.showMinutes })}
-          />
-          {' '}
-          Display minutes
-        </label>
-
-        <label>
-          <input
-            type="checkbox"
-            checked={this.props.showDate}
-            onChange={() => this.props.onChange({ showDate: ! this.props.showDate })}
-          />
-          {' '}
-          Display the date
-        </label>
-      </div>
-    );
-  }
-}
+    <label>
+      <input
+        type="checkbox"
+        checked={data.showDate}
+        onChange={() => setData({ ...data, showDate: !data.showDate })}
+      />{' '}
+      Display the date
+    </label>
+  </div>
+);
 
 export default TimeSettings;
