@@ -1,2 +1,13 @@
-export { default as Time } from './Time';
-export { default as TimeSettings } from './TimeSettings';
+import { Config } from '../../types';
+import Time from './Time';
+import TimeSettings from './TimeSettings';
+
+const config: Config = {
+  key: 'widget/time',
+  name: 'Time',
+  description: 'Be on time.',
+  dashboardComponent: Time,
+  settingsComponent: TimeSettings,
+};
+
+export default config;
