@@ -1,62 +1,72 @@
-import * as React from 'react';
-import { checkedIcon, uncheckedIcon, removeIcon } from '../../../app/ui';
-import { Todo } from './interfaces';
+import React, { FC, useLayoutEffect, useRef } from 'react';
+
+import { useKeyPress } from '../../../hooks';
+import { Icon, RemoveIcon } from '../../../views/shared';
+import { State } from './reducer';
 import './TodoItem.sass';
 
 interface Props {
-  item: Todo;
+  item: State[number];
   onToggle(): void;
   onUpdate(contents: string): void;
   onDelete(): void;
 }
 
-class TodoItem extends React.Component<Props> {
-  ref: HTMLSpanElement | null;
+const TodoItem: FC<Props> = ({ item, onDelete, onUpdate, onToggle }) => {
+  const ref = useRef<HTMLSpanElement>(null);
 
-  shouldComponentUpdate(nextProps: Props) {
-    if (! this.ref) {
-      return true;
+  useLayoutEffect(() => {
+    if (ref.current) {
+      ref.current.innerText = item.contents;
     }
+  }, [item.contents]);
 
-    return (
-      nextProps.item.id !== this.props.item.id ||
-      nextProps.item.completed !== this.props.item.completed ||
-      nextProps.item.contents !== this.ref.innerText
-    );
-  }
+  useKeyPress(
+    event => {
+      if (event.target === ref.current) {
+        event.preventDefault();
 
-  render() {
-    const { item, onDelete, onUpdate, onToggle } = this.props;
+        if (ref.current) {
+          ref.current.blur();
+        }
+      }
+    },
+    ['Enter'],
+    false,
+  );
 
-    return (
-      <div className="TodoItem">
-        <a onClick={onToggle}>
-          {item.completed ? checkedIcon : uncheckedIcon}
-        </a>
+  useKeyPress(
+    event => {
+      if (event.target === ref.current) {
+        event.preventDefault();
 
-        <span
-          ref={ref => this.ref = ref}
-          contentEditable={true}
-          onBlur={event => onUpdate(event.currentTarget.innerText)}
-          onInput={event => onUpdate(event.currentTarget.innerText)}
-          onKeyDownCapture={this.onKeyDown}
-          suppressContentEditableWarning={true}
-        >
-          {item.contents}
-        </span>
+        if (ref.current) {
+          // Reset contents on escape
+          ref.current.innerText = item.contents;
+          ref.current.blur();
+        }
+      }
+    },
+    ['Escape'],
+    false,
+  );
 
-        <a onClick={onDelete} className="delete">{removeIcon}</a>
-      </div>
-    );
-  }
+  return (
+    <div className="TodoItem">
+      <span
+        ref={ref}
+        contentEditable={true}
+        onBlur={event => onUpdate(event.currentTarget.innerText)}
+      />
 
-  private onKeyDown = (event: React.KeyboardEvent<HTMLSpanElement>) => {
-    // Did we press enter while editing?
-    if (event.keyCode === 13 && this.ref) {
-      event.preventDefault();
-      this.ref.blur();
-    }
-  }
-}
+      <a onMouseDown={onToggle} className="complete">
+        <Icon name={item.completed ? 'check-circle' : 'circle'} />
+      </a>
+      <a onMouseDown={onDelete} className="delete">
+        <RemoveIcon />
+      </a>
+    </div>
+  );
+};
 
 export default TodoItem;
