@@ -1,50 +1,21 @@
-import * as React from 'react';
-import { Settings } from './interfaces';
+import React, { FC } from 'react';
 
-interface Props extends Settings {
-  onChange: (settings: Partial<Settings>) => void;
-}
+import { Props, defaultData } from './types';
 
-const TodoSettings: React.StatelessComponent<Props> = ({
-  show = 3,
-  textAlign = 'left',
-  onChange,
-}) => (
+const TodoSettings: FC<Props> = ({ data = defaultData, setData }) => (
   <div className="SearchSettings">
-    <p>
-      <label>
-        Tasks to show
-        <input
-          type="number"
-          min="1"
-          onChange={event => onChange({ show: Number(event.target.value) })}
-          placeholder="Number of todo items to show"
-          value={show}
-        />
-      </label>
-    </p>
-
-    <p>
-      <label>
-        <input
-          type="radio"
-          checked={textAlign === 'inherit'}
-          onChange={() => onChange({ textAlign: 'inherit' })}
-        />
-
-        Default text align
-      </label>
-
-      <label>
-        <input
-          type="radio"
-          checked={textAlign === 'left'}
-          onChange={() => onChange({ textAlign: 'left' })}
-        />
-
-        Left text align
-      </label>
-    </p>
+    <label>
+      Tasks to show
+      <input
+        type="number"
+        min="0"
+        onChange={event =>
+          setData({ ...data, show: Number(event.target.value) })
+        }
+        placeholder="Number of todo items to show"
+        value={data.show}
+      />
+    </label>
   </div>
 );
 
