@@ -1,2 +1,13 @@
-export { default as Todo } from './Todo';
-export { default as TodoSettings } from './TodoSettings';
+import { Config } from '../../types';
+import Todo from './Todo';
+import TodoSettings from './TodoSettings';
+
+const config: Config = {
+  key: 'widget/todo',
+  name: 'Todos',
+  description: 'Add reminders to procrastinate.',
+  dashboardComponent: Todo,
+  settingsComponent: TodoSettings,
+};
+
+export default config;
