@@ -1,126 +1,129 @@
-import * as React from 'react';
+import React, { FC, FormEvent, useState } from 'react';
+
+import { useToggle } from '../../../hooks';
+import { Icon } from '../../../views/shared';
+import { geocodeLocation, getCurrentLocation } from './api';
+import { Coordinates } from './types';
 import './LocationInput.sass';
-const navigationIcon = require('feather-icons/dist/icons/navigation.svg');
-const searchIcon = require('feather-icons/dist/icons/search.svg');
 
-interface Props {
-  latitude: number|string;
-  longitude: number|string;
-  onChange: (location: { latitude?: number|string, longitude?: number|string }) => void;
-}
+type Props = {
+  latitude?: number;
+  longitude?: number;
+  onChange: (coords: Coordinates) => void;
+};
 
-interface State {
-  lookup: boolean;
-  query: string;
-}
+const GeocodeInput: FC<Props> = ({ onChange }) => {
+  const [query, setQuery] = useState('');
 
-class LocationInput extends React.PureComponent<Props, State> {
-  state: State = {
-    lookup: ! (this.props.latitude && this.props.longitude),
-    query: '',
+  const handleGeocode = (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+
+    geocodeLocation(query)
+      .then(onChange)
+      .catch(() => {
+        alert('Unable to find location. Please try again later.');
+      });
   };
 
-  render() {
-    // Lookup mode
-    if (this.state.lookup) {
-      return (
-        <div className="LocationInput">
-            <form onSubmit={event => { event.preventDefault(); this.geocodeLocation(); }}>
-              <div className="grid" style={{ gridTemplateColumns: '1fr auto' }}>
-                <label htmlFor="LocationInput__query">
-                  Search for city
-                </label>
+  return (
+    <form onSubmit={handleGeocode}>
+      <div className="grid" style={{ gridTemplateColumns: '1fr auto' }}>
+        <label htmlFor="LocationInput__query">Search for city</label>
 
-                <div />
+        <div />
 
-                <input
-                  id="LocationInput__query"
-                  placeholder="City or location"
-                  type="text"
-                  value={this.state.query}
-                  onChange={event => this.setState({ query: event.target.value })}
-                />
+        <input
+          id="LocationInput__query"
+          placeholder="City or location"
+          type="text"
+          value={query}
+          onChange={event => setQuery(event.target.value)}
+        />
 
-                <button type="submit" className="button--primary button--icon">
-                  <i dangerouslySetInnerHTML={{ __html: searchIcon }} />
-                </button>
-              </div>
-            </form>
-
-            <a onClick={() => this.setState({ lookup: false })}>
-              Enter coordinates
-            </a>
-        </div>
-      );
-    }
-
-    // Coordinates mode
-    const geolocationAvailable = 'geolocation' in navigator;
-
-    return (
-      <div className="LocationInput">
-        <div className="grid" style={{ gridTemplateColumns: geolocationAvailable ? '1fr 1fr auto' : '1fr 1fr' }}>
-          <label htmlFor="LocationInput__latitude">
-            Latitude
-          </label>
-
-          <label htmlFor="LocationInput__longitude">
-            Longitude
-          </label>
-
-          {geolocationAvailable && <div />}
-
-          <input
-            id="LocationInput__latitude"
-            type="text"
-            value={this.props.latitude}
-            onChange={event => this.props.onChange({ latitude: event.target.value })}
-          />
-
-          <input
-            id="LocationInput__longitude"
-            type="text"
-            value={this.props.longitude}
-            onChange={event => this.props.onChange({ longitude: event.target.value })}
-          />
-
-          {geolocationAvailable &&
-            <button className="button--primary button--icon" onClick={this.getCurrentLocation}>
-              <i dangerouslySetInnerHTML={{ __html: navigationIcon }} />
-            </button>
-          }
-        </div>
-
-        <a onClick={() => this.setState({ lookup: true })}>
-          Search for city
-        </a>
+        <button type="submit" className="button--primary button--icon">
+          <Icon name="search" />
+        </button>
       </div>
-    );
-  }
+    </form>
+  );
+};
 
-  private getCurrentLocation = () => {
-    navigator.geolocation.getCurrentPosition(
-      positon => this.props.onChange({
-        latitude: positon.coords.latitude,
-        longitude: positon.coords.longitude,
-      }),
-      error => alert('Unable to determine your location.'),
-    );
-  }
+const geolocationAvailable = 'geolocation' in navigator;
 
-  private geocodeLocation = () => {
-    fetch(`https://nominatim.openstreetmap.org/search.php?format=json&q=${this.state.query}`)
-      .then(res => res.json())
-      .then(data => this.props.onChange({
-        latitude: Number(data[0].lat).toFixed(4),
-        longitude: Number(data[0].lon).toFixed(4),
-      }))
-      .then(() => this.setState({ lookup: false, query: '' }))
-      .catch(() => {
-        alert('Unable to determine location. Please try again later.');
-        this.setState({ lookup: false });
-      });
-  }
-}
+const CoordinateInput: FC<Props> = ({ latitude, longitude, onChange }) => {
+  const handleLocate = () => {
+    getCurrentLocation()
+      .then(onChange)
+      .catch(err => alert(`Cannot determine your location: ${err.message}`));
+  };
+
+  return (
+    <div className="LocationInput">
+      <div
+        className="grid"
+        style={{
+          gridTemplateColumns: geolocationAvailable
+            ? '1fr 1fr auto'
+            : '1fr 1fr',
+        }}
+      >
+        <label htmlFor="LocationInput__latitude">Latitude</label>
+
+        <label htmlFor="LocationInput__longitude">Longitude</label>
+
+        {geolocationAvailable && <div />}
+
+        <input
+          id="LocationInput__latitude"
+          type="text"
+          value={latitude}
+          onChange={event => onChange({ latitude: Number(event.target.value) })}
+        />
+
+        <input
+          id="LocationInput__longitude"
+          type="text"
+          value={longitude}
+          onChange={event =>
+            onChange({ longitude: Number(event.target.value) })
+          }
+        />
+
+        {geolocationAvailable && (
+          <button
+            className="button--primary button--icon"
+            onClick={handleLocate}
+          >
+            <Icon name="navigation" />
+          </button>
+        )}
+      </div>
+    </div>
+  );
+};
+
+const LocationInput: FC<Props> = ({ onChange, ...props }) => {
+  const hasCoordinates = props.longitude && props.latitude;
+  const [lookUp, toggleLookUp] = useToggle(!hasCoordinates);
+
+  const handleChange = (coords: Coordinates) => {
+    onChange(coords);
+    if (lookUp) toggleLookUp();
+  };
+
+  return (
+    <div className="LocationInput">
+      {lookUp ? (
+        <GeocodeInput {...props} onChange={handleChange} />
+      ) : (
+        <CoordinateInput {...props} onChange={handleChange} />
+      )}
+
+      <a onClick={toggleLookUp}>
+        {lookUp ? 'Enter coordinates' : 'Search for city'}
+      </a>
+    </div>
+  );
+};
 
 export default LocationInput;
