@@ -1,13 +1,13 @@
 export const weatherIcons = {
-  'clear-day': require('feather-icons/dist/icons/sun.svg'),
-  'clear-night': require('feather-icons/dist/icons/moon.svg'),
-  'rain': require('feather-icons/dist/icons/cloud-rain.svg'),
-  'snow': require('feather-icons/dist/icons/cloud-snow.svg'),
-  'sleet': require('feather-icons/dist/icons/cloud-drizzle.svg'),
-  'wind': require('feather-icons/dist/icons/wind.svg'),
-  'fog': require('feather-icons/dist/icons/cloud-rain.svg'),
-  'cloudy': require('feather-icons/dist/icons/cloud.svg'),
-  'partly-cloudy-day': require('feather-icons/dist/icons/cloud.svg'),
-  'partly-cloudy-night': require('feather-icons/dist/icons/cloud.svg'),
-  'thunderstorm': require('feather-icons/dist/icons/cloud-lightning.svg'),
+  'clear-day': 'sun',
+  'clear-night': 'moon',
+  rain: 'cloud-rain',
+  snow: 'cloud-snow',
+  sleet: 'cloud-drizzle',
+  wind: 'wind',
+  fog: 'cloud-rain',
+  cloudy: 'cloud',
+  'partly-cloudy-day': 'cloud',
+  'partly-cloudy-night': 'cloud',
+  thunderstorm: 'cloud-lightning',
 };
