@@ -9,7 +9,7 @@
 // This link also includes instructions on opting out of this behavior.
 
 export function register() {
-  if (process.env.NODE_ENV === 'production' && 'serviceWorker' in navigator) {
+  if ('serviceWorker' in navigator) {
     window.addEventListener('load', () => {
       const swUrl = `/service-worker.js`;
       navigator.serviceWorker
@@ -18,7 +18,7 @@ export function register() {
           registration.onupdatefound = () => {
             const installingWorker = registration.installing;
             if (!installingWorker) {
-                return;
+              return;
             }
 
             installingWorker.onstatechange = () => {
@@ -28,19 +28,19 @@ export function register() {
                   // the fresh content will have been added to the cache.
                   // It's the perfect time to display a "New content is
                   // available; please refresh." message in your web app.
-                  console.log('New content is available; please refresh.'); // tslint:disable-line
+                  console.log('New content is available; please refresh.');
                 } else {
                   // At this point, everything has been precached.
                   // It's the perfect time to display a
                   // "Content is cached for offline use." message.
-                  console.log('Content is cached for offline use.'); // tslint:disable-line
+                  console.log('Content is cached for offline use.');
                 }
               }
             };
           };
         })
         .catch(error => {
-          console.error('Error during service worker registration:', error); // tslint:disable-line
+          console.error('Error during service worker registration:', error);
         });
     });
   }
